% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_disaggs}
\alias{check_disaggs}
\title{Validate disaggs for export}
\usage{
check_disaggs(df, ref_de)
}
\arguments{
\item{df}{HFR data frame containing reshaped submission}

\item{ref_de}{CIR Data Element Reference Data as data frame}
}
\value{
list of row ids with invalid indicators' disaggregations
}
\description{
Validate disaggs for export
}
