% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_numdenom}
\alias{check_numdenom}
\title{Validate Numerator/Denominator for export}
\usage{
check_numdenom(df)
}
\arguments{
\item{df}{HFR data frame containing reshaped submission}
}
\value{
list of row ids with invalid numerator/denominator values
}
\description{
Validate Numerator/Denominator for export
}
