% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_operatingunit}
\alias{check_operatingunit}
\title{Check OU}
\usage{
check_operatingunit(df, ou)
}
\arguments{
\item{df}{Data frame from transformed submission}

\item{ou}{Name of OU/Country submitting data}
}
\description{
Check OU
}
