% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_orgunituids}
\alias{check_orgunituids}
\title{Validate orgunituids for export}
\usage{
check_orgunituids(df, ref_orgs)
}
\arguments{
\item{df}{HFR data framed created by \code{cir_process_template()}}

\item{ref_orgs}{Datim OU Orgunits Reference Data as data frame}
}
\description{
Validate orgunituids for export
}
