% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_folder}
\alias{cir_folder}
\title{Get processing folder path}
\usage{
cir_folder(folderpath, type = "raw")
}
\arguments{
\item{folderpath}{Parent folder path}

\item{type}{Folder type}
}
\description{
Get processing folder path
}
