% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_gather.R
\name{cir_gather}
\alias{cir_gather}
\title{Reshape CIR Data frame long}
\usage{
cir_gather(df)
}
\arguments{
\item{df}{CIR data frame imported via \code{cir_import()}}
}
\description{
Reshape CIR Data frame long
}
