% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_import.R
\name{cir_import}
\alias{cir_import}
\title{Import template sheet(s)}
\usage{
cir_import(filepath, sheets = NULL, template = NULL)
}
\arguments{
\item{filepath}{filepath to submitted template}
}
\description{
Import template sheet(s)
}
\examples{
\dontrun{
#identify whether template is long or wide
  filepath <- "~/CIRG_FY22_Q1_Zimbabwe_20220211.xlsx"
  df_hfr <- cir_import(filepath) }
}
