% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_transform.R
\name{cir_join_meta}
\alias{cir_join_meta}
\title{Join Meta to Dataset}
\usage{
cir_join_meta(df_cir, df_meta)
}
\arguments{
\item{df_cir}{CIR data frame imported via \code{cir_import()}}

\item{df_meta}{CIR Meta data}
}
\description{
Join Meta to Dataset
}
