% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_output}
\alias{cir_output}
\title{Output intermediate data for files}
\usage{
cir_output(.df_out, .subm, .name, type = "metadata")
}
\arguments{
\item{.df_out}{Data to be outputs}

\item{.subm}{Submission file}

\item{.name}{Output base name - no extension needed, default is \code{.csv}}

\item{type}{Output data type, default is set to metadata}
}
\description{
Output intermediate data for files
}
