% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_processing_template.R
\name{cir_processing}
\alias{cir_processing}
\title{Process CIR Submissions}
\usage{
cir_processing(filepath)
}
\arguments{
\item{filepath}{Full filename}
}
\value{
CIR Submission as Tibble
}
\description{
Process CIR Submissions
}
