% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_transform.R
\name{cir_reshape}
\alias{cir_reshape}
\title{Transform CIR Data Frame}
\usage{
cir_reshape(df_cir, clean = TRUE)
}
\arguments{
\item{df_cir}{CIR raw and restricted data frame}

\item{clean}{Clean indicators and disaggregates - default is TRUE}
}
\value{
Cleaned and tidy data frame
}
\description{
Transform CIR Data Frame
}
\note{
This can also be accomplished with \code{cir_gather()} + \code{cir_munge_string()}
}
