% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_reshape_checks}
\alias{cir_reshape_checks}
\title{Reshape Validation Checks data frame}
\usage{
cir_reshape_checks(.df_checks, vname = "location")
}
\arguments{
\item{.df_checks}{CIR Processing checks as data frame}

\item{vname}{Column name to be used for values_to parameter of \code{pivot_longer()}}
}
\value{
Validation Checks in long format
}
\description{
Reshape Validation Checks data frame
}
