% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_restrict_cols.R
\name{cir_restrict_cols}
\alias{cir_restrict_cols}
\title{Restrict CIR data frame columns}
\usage{
cir_restrict_cols(df)
}
\arguments{
\item{df}{CIR data frame imported via \code{cir_import()}}
}
\description{
Restrict CIR data frame columns
}
