% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_store_meta.R
\name{cir_store_meta}
\alias{cir_store_meta}
\title{Store Meta Data Information about Template}
\usage{
cir_store_meta(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
store meta data information for later validations
}
