% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_template_cols}
\alias{cir_template_cols}
\title{Validate template}
\usage{
cir_template_cols(df_cir, template = "long")
}
\arguments{
\item{template}{Template name to validate against}

\item{df}{Data frame to check against}
}
\value{
List of required column names
}
\description{
Validate template
}
\examples{
\dontrun{
tmp = "Semi-wide"
tmp_cols <- cir_template_cols(df, template = tmp)}
}
