% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_template_ta}
\alias{cir_template_ta}
\title{Extract Technical Areas}
\usage{
cir_template_ta(df_cir)
}
\arguments{
\item{df_cir}{Data frame to extract columns from}
}
\value{
Technical Area name(s)
}
\description{
Extract Technical Areas
}
\examples{
\dontrun{
ta <- cir_template_ta(df)}
}
