% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_vsheets}
\alias{cir_vsheets}
\title{Get list of visible excel sheets}
\usage{
cir_vsheets(.subm, return_names = FALSE)
}
\arguments{
\item{.subm}{Submission file}

\item{return_names}{Should the function return the list of valid data sheet names? Default is false}
}
\value{
Worksheet visibility as data frame
}
\description{
Get list of visible excel sheets
}
