% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{clean_orgview}
\alias{clean_orgview}
\title{Clean OrgUnits SQLView}
\usage{
clean_orgview(.data, levels)
}
\arguments{
\item{.data}{Data Frame}

\item{levels}{Org Levels}
}
\description{
Clean OrgUnits SQLView
}
