% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{datim_mechs}
\alias{datim_mechs}
\title{Pull Orgunits}
\usage{
datim_mechs(username, password, cntry, base_url = NULL)
}
\arguments{
\item{username}{Datim username}

\item{password}{Datim password}

\item{cntry}{Country name}

\item{base_url}{Datim API Base URL}
}
\value{
OU Mechanisms as data frame
}
\description{
Pull Orgunits
}
