% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{datim_orglevels}
\alias{datim_orglevels}
\title{Country organisation units levels}
\usage{
datim_orglevels(cntry)
}
\arguments{
\item{cntry}{Country name}
}
\value{
Country org levels as data frame
}
\description{
Country organisation units levels
}
