% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{empty_to_chr}
\alias{empty_to_chr}
\title{Return none if vector/list is empty}
\usage{
empty_to_chr(obj, type = "atomic")
}
\arguments{
\item{obj}{text to be printed}

\item{type}{Type of the object, default set to atomic (vector or list). Options are atomic or string}
}
\description{
Return none if vector/list is empty
}
