% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_missing}
\alias{get_missing}
\title{Get rows with missing values}
\usage{
get_missing(df, var)
}
\arguments{
\item{df}{data frame}

\item{var}{variable to count missing values}
}
\description{
Get rows with missing values
}
