% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_initial.R
\name{has_cirgtab}
\alias{has_cirgtab}
\title{Determine if there are tabs to import}
\usage{
has_cirgtab(filepath)
}
\arguments{
\item{filepath}{submission file path}
}
\description{
Determine if there are tabs to import
}
\note{
TODO - Change function name to \code{has_cirtab}
}
