% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_initial.R
\name{has_metatab}
\alias{has_metatab}
\title{Determine whether meta tab exists}
\usage{
has_metatab(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Determine whether meta tab exists
}
\note{
TODO - Change function name to \code{has_metatab}
}
