% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interesting-package.R
\docType{package}
\name{Interesting-package}
\alias{Interesting}
\alias{Interesting-package}
\title{Interesting: Interesting}
\description{
Processing and validation scripts for CIRG Quarterly Data
}
\author{
\strong{Maintainer}: Karishma Srikanth \email{ksrikanth@usaid.gov}

Authors:
\itemize{
  \item Nelly Maina \email{nmaina@usaid.gov}
  \item Nashiva McDavid \email{nmcdavid@usaid.gov}
  \item Baboyma Kagniniwa \email{bkagniniwa@usaid.gov}
}

}
\keyword{internal}
