% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_value}
\alias{match_value}
\title{Get rows with values not matching a pattern}
\usage{
match_value(df, var, pattern = "FY\\\\d{2}Q\\\\d{1}")
}
\arguments{
\item{df}{data frame}

\item{var}{variable to count missing values}
}
\description{
Get rows with values not matching a pattern
}
