% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paint_ifempty}
\alias{paint_ifempty}
\title{Paint if empty}
\usage{
paint_ifempty(txt, true_paint = crayon::blue, false_paint = crayon::yellow)
}
\arguments{
\item{txt}{text to be painted and printed}

\item{true_paint}{crayon function to execute if ~is_empty(txt) or txt == ""}

\item{false_patin}{crayon function to execute}
}
\description{
Paint if empty
}
