% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paint_ifna}
\alias{paint_ifna}
\title{Paint if na}
\usage{
paint_ifna(txt, true_paint = crayon::yellow, false_paint = crayon::blue)
}
\arguments{
\item{txt}{text to be painted and printed}

\item{true_paint}{crayon function to execute if ~is.na(txt)}
}
\description{
Paint if na
}
