% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paint_ifnull}
\alias{paint_ifnull}
\title{Paint if null}
\usage{
paint_ifnull(obj, true_paint = crayon::red, false_paint = crayon::blue)
}
\arguments{
\item{obj}{text to be painted and printed}

<<<<<<< HEAD
\item{true_paint}{crayon function to execute if is.null(obj)}

\item{false_paint}{crayon function to execute if !is.null(ojb)}
=======
\item{true_paint}{crayon function to execute if ~is.na(txt)}
>>>>>>> feature/process-flow
}
\description{
Paint if null
}
