% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paint_iftrue}
\alias{paint_iftrue}
\title{Paint if true}
\usage{
paint_iftrue(value, true_paint = crayon::green, false_paint = crayon::red)
}
\arguments{
\item{value}{text to be painted and printed}

\item{true_paint}{crayon function to execute if ~is.na(txt)}
}
\description{
Paint if true
}
