% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{rename_orgview}
\alias{rename_orgview}
\title{Rename OrgUnit SQLView}
\usage{
rename_orgview(.data, levels)
}
\arguments{
\item{.data}{Data Frame}

\item{levels}{Org Levels}
}
\description{
Rename OrgUnit SQLView
}
