% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{reshape_orgview}
\alias{reshape_orgview}
\title{Reshape OrgUnits SQLView}
\usage{
reshape_orgview(.data, levels)
}
\arguments{
\item{.data}{Data Frame}

\item{levels}{Org Levels}
}
\description{
Reshape OrgUnits SQLView
}
