% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_cols_long}
\alias{template_cols_long}
\title{Column names for FY22 CIRG Long Template}
\format{
A data frame with 0 rows and 14 variables:
}
\usage{
template_cols_long
}
\description{
A dataset with the column names for the CIRG FY22 Templat in long format
}
\keyword{datasets}
