% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mech.R
\name{update_meta_mechs}
\alias{update_meta_mechs}
\title{Update USAID Mechanism meta table}
\usage{
update_meta_mechs(savefolder = "out/DATIM", upload = FALSE)
}
\arguments{
\item{savefolder}{folderpath to save, default = "out/DATIM"}

\item{upload}{should the new table be pushed to Google Drive and s3? default = FALSE}
}
\description{
Update USAID Mechanism meta table
}
