% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_import.R
\name{validate_import}
\alias{validate_import}
\title{Validation on import}
\usage{
validate_import(df_cir, template)
}
\arguments{
\item{df_cir}{df created during \code{cir_import()}}

\item{template}{submission template}
}
\description{
Validation on import
}
