% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{validate_output}
\alias{validate_output}
\title{Validation Checks}
\usage{
validate_output(df, refs, content = FALSE)
}
\arguments{
\item{df}{HFR data framed created by \code{cir_process_template()}}

\item{refs}{Reference datasets to include orgs, mechs, and data elements}

\item{content}{check full dataset}
}
\value{
list containing validated data along with results of the checks
}
\description{
Validation Checks
}
