% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_matches}
\alias{var_matches}
\title{Check if any variable matches this pattern}
\usage{
var_matches(df, pattern)
}
\arguments{
\item{df}{data frame to check against}

\item{pattern}{quoted variable of interest}
}
\description{
Check if any variable matches this pattern
}
