% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_orgunituids}
\alias{check_orgunituids}
\title{Validate orgunituids}
\usage{
check_orgunituids(df, ref_orgs)
}
\arguments{
\item{df}{HFR data frame containing reshaped submission}

\item{ref_orgs}{Datim OU Orgunits Reference Data as data frame}
}
\value{
list of row ids with invalid orgunituid
}
\description{
Validate orgunituids
}
