% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_folder}
\alias{cir_folder}
\title{Get processing folder path}
\usage{
cir_folder(type = "raw", dt = NULL)
}
\arguments{
\item{type}{Folder type}

\item{dt}{porcessing date}
}
\description{
Get processing folder path
}
