% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_transform.R
\name{cir_gather}
\alias{cir_gather}
\title{Reshape CIR Data frame to long}
\usage{
cir_gather(df_cir)
}
\arguments{
\item{df_cir}{CIR data frame imported via \code{cir_import()}}
}
\description{
Reshape CIR Data frame to long
}
