% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_munge_string.R
\name{cir_munge_string}
\alias{cir_munge_string}
\title{clean up/standardize string text for indicators and disaggs}
\usage{
cir_munge_string(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{cir_import()}}
}
\description{
clean up/standardize string text for indicators and disaggs
}
