% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_processing_template.R
\name{cir_processing}
\alias{cir_processing}
\title{Process CIR Submissions}
\usage{
cir_processing(filepath, archive = FALSE)
}
\arguments{
\item{filepath}{Full path of filename}

\item{archive}{Should the file be archvied if successfully processed? Default is false}
}
\value{
CIR Submission as Tibble
}
\description{
Process CIR Submissions
}
