% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_import.R
\name{cir_reshape}
\alias{cir_reshape}
\title{Transform CIR Data Frame}
\usage{
cir_reshape(df)
}
\arguments{
\item{df}{CIR raw and restricted data frame}
}
\value{
Cleaned and tidy data frame
}
\description{
Transform CIR Data Frame
}
\note{
This can also be accomplished with \code{cir_gather()} + \code{cir_munge_string()}
}
