% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cir_setup}
\alias{cir_setup}
\title{Setup processing folders}
\usage{
cir_setup(folder = "cirg-submissions", dt = NULL)
}
\arguments{
\item{folder}{Folder path used for CIRG Data processing}

\item{dt}{Processing date. If null, current date is used}
}
\description{
Setup processing folders
}
