% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{datim_mechs}
\alias{datim_mechs}
\title{Pull Orgunits}
\usage{
datim_mechs(username, password, cntry, base_url = NULL)
}
\arguments{
\item{username}{}
}
\value{
OU Orgunit as data frame
}
\description{
Pull Orgunits
}
