% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{datim_sqlviews}
\alias{datim_sqlviews}
\title{Query Datim SQLViews}
\usage{
datim_sqlviews(
  username,
  password,
  view_name = NULL,
  dataset = FALSE,
  datauid = NULL,
  query = NULL,
  base_url = NULL
)
}
\arguments{
\item{username}{Datim username}

\item{password}{Datim password}

\item{view_name}{Datim SQLView name}

\item{dataset}{Return SQLView dataset or uid? Default is false}

\item{datauid}{Data UID}

\item{query}{SQLView Query params, a list containing type and params key value pairs}

\item{base_url}{Datim API Base URL}
}
\value{
SQLView uid or dataset as data frame
}
\description{
Query Datim SQLViews
}
