% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_queries.R
\name{datim_sqlviews}
\alias{datim_sqlviews}
\title{Query Datim SQLViews}
\usage{
datim_sqlviews(
  username,
  password,
  view_name = NULL,
  dataset = FALSE,
  datauid = NULL,
  query = NULL,
  base_url = NULL
)
}
\arguments{
\item{password}{}
}
\description{
Query Datim SQLViews
}
