% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{empty_to_chr}
\alias{empty_to_chr}
\title{Return none if vector/list is empty}
\usage{
empty_to_chr(obj)
}
\arguments{
\item{obj}{text to be printed}
}
\description{
Return none if vector/list is empty
}
